tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/animation-duration_one_value.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - animation-duration value with one list entry</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#the-animation-duration-property-">
   <meta name="flags" content="dom">
   <meta name="assert" content="1000ms is a valid animation-duration value">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes darken {
         from {
            background-color: blue;
         }
         to {
            background-color: black;
         }
      }
      div {
         width: 100px;
         height: 100px;
         background: blue;
      }
      .test {
         animation-name: darken;
         animation-duration: 1000ms;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelector('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         function aboutEqual(observed, actual) {
            var maxFactor = 0.01;
            return (Math.abs(observed - actual) < maxFactor * actual);
         }
         var timeout = setTimeout(function() {
            test(function(){assert_true(false);}, testName); done();
            log.data = 'FAIL';
         }, 1500);
         div.addEventListener('animationend', function(evt) {
            clearTimeout(timeout);
            var passed = aboutEqual(evt.elapsedTime, 1);
            test(function(){assert_true(passed);}, testName); done();
            log.data = (passed ? 'PASS if the animation occurred.' : 'FAIL');
         });
         div.className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The square below should gradually turn black and then instantly turn blue
   again.</p>
   <p>Testing...</p>
   <div></div>
</body>
</html>
